Reworking how factory stats are handled
Isolating the `MessageProcessingStats` type to it's own module.

Also exposes the factory hash operation to be public.
